Master <= Develop

+++ ParleoBackend/Controllers/AccountController.cs
@@ -22,6 +22,7 @@ using Microsoft.AspNetCore.Http;
 using System.IO;
 using Parleo.BLL.Extensions;
 using ParleoBackend.Validators.Common;
+using ParleoBackend.Services;
 
 namespace ParleoBackend.Controllers
 {
@@ -33,6 +34,7 @@ namespace ParleoBackend.Controllers
         private readonly IEmailService _emailService;
         private readonly IJwtService _jwtService;
         private readonly IImageSettings _accountImageSettings;
+        private readonly IJwtSettings _jwtSettings;
         private readonly IUtilityService _utilityService;
         private readonly IMapper _mapper;
         private readonly ILogger _logger;
@@ -44,7 +46,8 @@ namespace ParleoBackend.Controllers
             IEmailService emailService,
             ILogger<AccountController> logger,
             IImageSettings accountImageSettings,
-            IUtilityService utilityService
+            IUtilityService utilityService,
+            IJwtSettings jwtSettings
         )
         {
             _accountService = accountService;
@@ -54,10 +57,10 @@ namespace ParleoBackend.Controllers
             _emailService = emailService;
             _accountImageSettings = accountImageSettings;
             _utilityService = utilityService;
+            _jwtSettings = jwtSettings;
         }
 
         [HttpGet]
-        [Authorize]
         [ProducesResponseType((int)HttpStatusCode.NotFound)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> GetUsersPageAsync(
@@ -88,6 +91,7 @@ namespace ParleoBackend.Controllers
         }
 
         [HttpPost("register")]
+        [AllowAnonymous]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> RegisterAsync(UserRegistrationViewModel registrationViewModel)
         {
@@ -105,7 +109,7 @@ namespace ParleoBackend.Controllers
                 return BadRequest(new ErrorResponseFormat(Constants.Errors.USER_CREATION_FAILED));
             }
 
-            string tokenString = _jwtService.GetJWTToken(user);
+            string tokenString = _jwtService.GetJWTToken(user, new EmailClaimsService(_jwtSettings));
             await _accountService.AddAccountTokenAsync(
                 new AccountTokenModel()
                 {
@@ -119,6 +123,7 @@ namespace ParleoBackend.Controllers
         }
 
         [HttpPost("login")]
+        [AllowAnonymous]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> LoginAsync(UserLoginViewModel loginViewModel)
         {
@@ -140,19 +145,23 @@ namespace ParleoBackend.Controllers
             {
                 return BadRequest(new ErrorResponseFormat(Constants.Errors.INVALID_PASSWORD));
             }
-            string tokenString = _jwtService.GetJWTToken(user);
+            string tokenString = _jwtService.GetJWTToken(user, new ClaimsService(_jwtSettings));
 
             return Ok(new {token = tokenString});
         }
 
-        [HttpPut("{userId}")]
-        [Authorize]
+        [HttpPut("current")]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
         [ProducesResponseType((int)HttpStatusCode.NoContent)]
         public async Task<IActionResult> EditAsync(
-            Guid userId,
             [FromBody] UpdateUserViewModel entity)
         {
+            string id = User.FindFirst(JwtRegisteredClaimNames.Jti).Value;
+            if (!Guid.TryParse(id, out Guid userGuid))
+            {
+                return BadRequest(new ErrorResponseFormat(Constants.Errors.WRONG_GUID_FORMAT));
+            }
+
             var validator = new UpdateUserViewModelValidator(_utilityService, _mapper);
             ValidationResult result = validator.Validate(entity);
 
@@ -162,7 +171,7 @@ namespace ParleoBackend.Controllers
             }
 
             bool isEdited = await _accountService.UpdateUserAsync(
-                    userId, _mapper.Map<UpdateUserModel>(entity));
+                    userGuid, _mapper.Map<UpdateUserModel>(entity));
 
             if (!isEdited)
             {
@@ -173,17 +182,10 @@ namespace ParleoBackend.Controllers
         }
 
         [HttpGet("{userId}")]
-        [Authorize]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public async Task<IActionResult> GetUserByIdAsync(string userId)
+        public async Task<IActionResult> GetUserByIdAsync(Guid userId)
         {
-            Guid userGuid;
-            if (!Guid.TryParse(userId, out userGuid))
-            {
-                return BadRequest(new ErrorResponseFormat(Constants.Errors.WRONG_GUID_FORMAT));
-            }
-
-            UserModel user = await _accountService.GetUserByIdAsync(new Guid(userId));
+            UserModel user = await _accountService.GetUserByIdAsync(userId);
             if (user == null)
             {
                 return BadRequest(new ErrorResponseFormat(Constants.Errors.USER_NOT_FOUND));
@@ -197,7 +199,6 @@ namespace ParleoBackend.Controllers
         }
 
         [HttpGet("me")]
-        [Authorize]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> GetUserByTokenAsync()
         {
@@ -217,6 +218,7 @@ namespace ParleoBackend.Controllers
 
 
         [HttpGet("activate")]
+        [AllowAnonymous]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> GetActivatedUserAccount(string token)
         {
@@ -242,12 +244,11 @@ namespace ParleoBackend.Controllers
 
             return Ok(new {
                 id = user.Id,
-                token = _jwtService.GetJWTToken(user)
+                token = _jwtService.GetJWTToken(user, new ClaimsService(_jwtSettings))
             });
         }
 
-        [HttpPut("{userId}/image")]
-        [Authorize]
+        [HttpPut("current/image")]
         public async Task<IActionResult> AddUserAccountImage(IFormCollection formData)
         {
             if (formData == null)
@@ -262,10 +263,15 @@ namespace ParleoBackend.Controllers
                 return BadRequest();
             }
 
-            string accountImagePath = _accountImageSettings.AccountDestPath;
-            Guid userId = new Guid(User.FindFirst(JwtRegisteredClaimNames.Jti).Value);
-            UserModel user = await _accountService.GetUserByIdAsync(userId);
+            string id = User.FindFirst(JwtRegisteredClaimNames.Jti).Value;
+            if (!Guid.TryParse(id, out Guid userGuid))
+            {
+                return BadRequest(new ErrorResponseFormat(Constants.Errors.USER_NOT_FOUND));
+            }
 
+            UserModel user = await _accountService.GetUserByIdAsync(userGuid);
+
+            string accountImagePath = _accountImageSettings.AccountDestPath;
             if (user.AccountImage != null)
             {
                 System.IO.File.Delete(Path.Combine(accountImagePath, user.AccountImage));
@@ -275,16 +281,21 @@ namespace ParleoBackend.Controllers
 
             await _accountService.InsertUserAccountImageAsync(
                 accountImageUniqueName,
-                userId
+                userGuid
             );
 
             return Ok();
         }
 
-        [HttpPut("{userId}/location")]
-        [Authorize]
-        public async Task<IActionResult> UpdateUserLocation(Guid userId, [FromBody] LocationViewModel location)
+        [HttpPut("current/location")]
+        public async Task<IActionResult> UpdateUserLocation([FromBody] LocationViewModel location)
         {
+            string id = User.FindFirst(JwtRegisteredClaimNames.Jti).Value;
+            if (!Guid.TryParse(id, out Guid userGuid))
+            {
+                return BadRequest(new ErrorResponseFormat(Constants.Errors.USER_NOT_FOUND));
+            }
+
             var validator = new LocationViewModelValidator();
             ValidationResult result = validator.Validate(location);
             if (!result.IsValid)
@@ -292,12 +303,7 @@ namespace ParleoBackend.Controllers
                 return BadRequest(new ErrorResponseFormat(result.Errors.First().ErrorMessage));
             }
 
-            if (userId == null)
-            {
-                return BadRequest(new ErrorResponseFormat(Constants.Errors.USER_NOT_FOUND));
-            }
-
-            bool isEdited = await _accountService.UpdateUserLocationAsync(userId, _mapper.Map<LocationModel>(location));
+            bool isEdited = await _accountService.UpdateUserLocationAsync(userGuid, _mapper.Map<LocationModel>(location));
 
             if (!isEdited)
             {
